refactor: StudyMemberAchievementDto 분리 및 side effect 수정
