+++ src/main/java/com/join/core/avatar/dto/response/MyManagedStudyInfoResponse.java
@@ -1,8 +1,8 @@
 package com.join.core.avatar.dto.response;
 
-import com.join.core.avatar.domain.Avatar;
 import com.join.core.study.constant.StudyStatus;
 import com.join.core.study.domain.Study;
+import com.join.core.study.service.dto.StudyMemberAchievementDto;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 import java.util.List;
@@ -23,30 +23,6 @@ public record MyManagedStudyInfoResponse(
         @Schema(description = "스터디 카카오톡 링크", example = "https://open.kakao.com/o/joinjoinjoin")
         String kakaoUrl
 ) {
-    public record StudyMemberAchievementDto(
-            @Schema(description = "스터디원 아바타 토큰", example = "token")
-            String avatarToken,
-            @Schema(description = "스터디원 별명", example = "닉네임")
-            String nickname,
-            @Schema(description = "이 스터디에서의 평균 출석률", example = "87.5")
-            double averageAttendanceRate,
-            @Schema(description = "이 스터디에서의 평균 인증률", example = "87.5")
-            double averageProofRate,
-            @Schema(description = "이 스터디의 모든 인증이 승인되었는지에 대한 여부", example = "true")
-            boolean isFullyApproved
-    ) {
-
-        public static StudyMemberAchievementDto of(Avatar avatar, double averageAttendanceRate, double averageProofRate, boolean isFullyVerified) {
-            return new StudyMemberAchievementDto(
-                    avatar.getAvatarToken(),
-                    avatar.getNickname(),
-                    averageAttendanceRate,
-                    averageProofRate,
-                    isFullyVerified
-            );
-        }
-    }
-
     public static MyManagedStudyInfoResponse of(Study study, double teamAverageAttendanceRate, double teamAverageProofRate, List<StudyMemberAchievementDto> achievementDtos) {
         return new MyManagedStudyInfoResponse(
                 study.getStudyToken(),
